server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.111.0
uvicorn>=0.30.0
pytest>=8.0.0
openai>=1.76.0
